+++ app/src/main/java/com/example/bubbel/repository/BackendRepository.kt
@@ -1,78 +1,18 @@
 package com.example.bubbel.repository
 
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
+import android.content.Context
 import com.example.bubbel.model.backend.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.POST
 
-interface backendService {
-    @POST("/api/create_user")
-    fun createUser(@Body userData: InCreateUser): Call<ResCreateUser>
-    @POST("/api/auth_user")
-    fun authUser(@Body userData: InAuthUser): Call<ResAuthUser>
-    @POST("/api/deauth_user")
-    fun deauthUser(@Body userData: InDeauthUser): Call<ResDeauthUser>
-    @POST("/api/verify_account")
-    fun verifyAccount(@Body userData: InVerifyAccount): Call<ResVerifyAccount>
-    @POST("/api/send_verify")
-    fun sendVerify(@Body userData: InSendVerify): Call<ResSendVerify>
-    @POST("/api/set_user_profile")
-    fun setUserProfile(@Body userData: InSetUserProfile): Call<ResSetUserProfile>
-    @POST("/api/get_user_profile")
-    fun getUserProfile(@Body userData: InGetUserProfile): Call<ResGetUserProfile>
-    @POST("/api/delete_user")
-    fun deleteUser(@Body userData: InDeleteUser): Call<ResDeleteUser>
-    @POST("/api/create_club")
-    fun createClub(@Body userData: InCreateClub): Call<ResCreateClub>
-    @POST("/api/get_club_profile")
-    fun getClubProfile(@Body userData: InGetClubProfile): Call<ResGetClubProfile>
-    @POST("/api/set_club_profile")
-    fun setClubProfile(@Body userData: InSetClubProfile): Call<ResSetClubProfile>
-    @POST("/api/delete_club")
-    fun deleteClub(@Body userData: InDeleteClub): Call<ResDeleteClub>
-    @POST("/api/get_user_profile_with_username")
-    fun getUserProfileWithUsername(@Body userData: InGetUserProfileWithUsername): Call<ResGetUserProfileWithUsername>
-    @POST("/api/add_friend_connection")
-    fun addFriendConnection(@Body userData: InAddFriendConnection): Call<ResAddFriendConnection>
-    @POST("/api/get_friend_connections")
-    fun getFriendConnections(@Body userData: InGetFriendConnections): Call<ResGetFriendConnections>
-    @POST("/api/remove_friend")
-    fun removeFriend(@Body userData: InRemoveFriend): Call<ResRemoveFriend>
-    @POST("/api/join_club")
-    fun joinClub(@Body userData: InJoinClub): Call<ResJoinClub>
-    @POST("/api/unjoin_club")
-    fun unjoinClub(@Body userData: InUnjoinClub): Call<ResUnjoinClub>
-    @POST("/api/get_club_members")
-    fun getClubMembers(@Body userData: InGetClubMembers): Call<ResGetClubMembers>
-    @POST("/api/get_user_clubs")
-    fun getUserClubs(@Body userData: InGetUserClubs): Call<ResGetUserClubs>
-    @POST("/api/regex_search_clubs")
-    fun regexSearchClubs(@Body userData: InRegexSearchClubs): Call<ResRegexSearchClubs>
-    @POST("/api/regex_search_users")
-    fun regexSearchUsers(@Body userData: InRegexSearchUsers): Call<ResRegexSearchUsers>
-}
 
-//  This was originally went in "RetrofitClient.kt"
-object RetrofitClient {
-    val api: backendService by lazy {
-        Retrofit.Builder()
-            .baseUrl("https://api.joinbubbel.com")
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(backendService::class.java)
-    }
-}
+class BackendRepository (context: Context) {
 
-//  This was originally went in "XXXRepository.kt"
-class BackendRepository {
-    private val backendService = RetrofitClient.api
+    private val retrofitClient = RetrofitClient(context).api.create(BackendService::class.java)
 
     suspend fun createUser(request: InCreateUser, onSuccess: (ResCreateUser?) -> Unit, onError: (String) -> Unit){
-        backendService.createUser(request).enqueue(object : Callback<ResCreateUser> {
+        retrofitClient.createUser(request).enqueue(object : Callback<ResCreateUser> {
             override fun onResponse(call: Call<ResCreateUser>, response: Response<ResCreateUser>) {
                 if (response.isSuccessful) {
                     val out: ResCreateUser? = response.body()
@@ -88,7 +28,7 @@ class BackendRepository {
         })
     }
     suspend fun authUser(request: InAuthUser,  onSuccess: (ResAuthUser?) -> Unit, onError: (String) -> Unit){
-        backendService.authUser(request).enqueue(object : Callback<ResAuthUser> {
+        retrofitClient.authUser(request).enqueue(object : Callback<ResAuthUser> {
             override fun onResponse(call: Call<ResAuthUser>, response: Response<ResAuthUser>) {
                 if (response.isSuccessful) {
                     val out: ResAuthUser? = response.body()
@@ -104,7 +44,7 @@ class BackendRepository {
         })
     }
     suspend fun deauthUser(request: InDeauthUser,  onSuccess: (ResDeauthUser?) -> Unit, onError: (String) -> Unit){
-        backendService.deauthUser(request).enqueue(object : Callback<ResDeauthUser> {
+        retrofitClient.deauthUser(request).enqueue(object : Callback<ResDeauthUser> {
             override fun onResponse(call: Call<ResDeauthUser>, response: Response<ResDeauthUser>) {
                 if (response.isSuccessful) {
                     val out: ResDeauthUser? = response.body()
@@ -120,7 +60,7 @@ class BackendRepository {
         })
     }
     suspend fun verifyAccount(request: InVerifyAccount,  onSuccess: (ResVerifyAccount?) -> Unit, onError: (String) -> Unit){
-        backendService.verifyAccount(request).enqueue(object : Callback<ResVerifyAccount> {
+        retrofitClient.verifyAccount(request).enqueue(object : Callback<ResVerifyAccount> {
             override fun onResponse(call: Call<ResVerifyAccount>, response: Response<ResVerifyAccount>) {
                 if (response.isSuccessful) {
                     val out: ResVerifyAccount? = response.body()
@@ -135,7 +75,7 @@ class BackendRepository {
         })
     }
     suspend fun sendVerify(request: InSendVerify,  onSuccess: (ResSendVerify?) -> Unit, onError: (String) -> Unit){
-        backendService.sendVerify(request).enqueue(object : Callback<ResSendVerify> {
+        retrofitClient.sendVerify(request).enqueue(object : Callback<ResSendVerify> {
             override fun onResponse(call: Call<ResSendVerify>, response: Response<ResSendVerify>) {
                 if (response.isSuccessful) {
                     val out: ResSendVerify? = response.body()
@@ -151,7 +91,7 @@ class BackendRepository {
         })
     }
     suspend fun setUserProfile(request: InSetUserProfile,  onSuccess: (ResSetUserProfile?) -> Unit, onError: (String) -> Unit){
-        backendService.setUserProfile(request).enqueue(object : Callback<ResSetUserProfile> {
+        retrofitClient.setUserProfile(request).enqueue(object : Callback<ResSetUserProfile> {
             override fun onResponse(call: Call<ResSetUserProfile>, response: Response<ResSetUserProfile>) {
                 if (response.isSuccessful) {
                     val out: ResSetUserProfile? = response.body()
@@ -167,7 +107,7 @@ class BackendRepository {
         })
     }
     suspend fun getUserProfile(request: InGetUserProfile,  onSuccess: (ResGetUserProfile?) -> Unit, onError: (String) -> Unit){
-        backendService.getUserProfile(request).enqueue(object : Callback<ResGetUserProfile> {
+        retrofitClient.getUserProfile(request).enqueue(object : Callback<ResGetUserProfile> {
             override fun onResponse(call: Call<ResGetUserProfile>, response: Response<ResGetUserProfile>) {
                 if (response.isSuccessful) {
                     val out: ResGetUserProfile? = response.body()
@@ -183,7 +123,7 @@ class BackendRepository {
         })
     }
     suspend fun deleteUser(request: InDeleteUser,  onSuccess: (ResDeleteUser?) -> Unit, onError: (String) -> Unit){
-        backendService.deleteUser(request).enqueue(object : Callback<ResDeleteUser> {
+        retrofitClient.deleteUser(request).enqueue(object : Callback<ResDeleteUser> {
             override fun onResponse(call: Call<ResDeleteUser>, response: Response<ResDeleteUser>) {
                 if (response.isSuccessful) {
                     val out: ResDeleteUser? = response.body()
@@ -199,7 +139,7 @@ class BackendRepository {
         })
     }
     suspend fun createClub(request: InCreateClub,  onSuccess: (ResCreateClub?) -> Unit, onError: (String) -> Unit){
-        backendService.createClub(request).enqueue(object : Callback<ResCreateClub> {
+        retrofitClient.createClub(request).enqueue(object : Callback<ResCreateClub> {
             override fun onResponse(call: Call<ResCreateClub>, response: Response<ResCreateClub>) {
                 if (response.isSuccessful) {
                     val out: ResCreateClub? = response.body()
@@ -215,7 +155,7 @@ class BackendRepository {
         })
     }
     suspend fun getClubProfile(request: InGetClubProfile,  onSuccess: (ResGetClubProfile?) -> Unit, onError: (String) -> Unit){
-        backendService.getClubProfile(request).enqueue(object : Callback<ResGetClubProfile> {
+        retrofitClient.getClubProfile(request).enqueue(object : Callback<ResGetClubProfile> {
             override fun onResponse(call: Call<ResGetClubProfile>, response: Response<ResGetClubProfile>) {
                 if (response.isSuccessful) {
                     val out: ResGetClubProfile? = response.body()
@@ -231,7 +171,7 @@ class BackendRepository {
         })
     }
     suspend fun setClubProfile(request: InSetClubProfile,  onSuccess: (ResSetClubProfile?) -> Unit, onError: (String) -> Unit){
-        backendService.setClubProfile(request).enqueue(object : Callback<ResSetClubProfile> {
+        retrofitClient.setClubProfile(request).enqueue(object : Callback<ResSetClubProfile> {
             override fun onResponse(call: Call<ResSetClubProfile>, response: Response<ResSetClubProfile>) {
                 if (response.isSuccessful) {
                     val out: ResSetClubProfile? = response.body()
@@ -247,7 +187,7 @@ class BackendRepository {
         })
     }
     suspend fun deleteClub(request: InDeleteClub,  onSuccess: (ResDeleteClub?) -> Unit, onError: (String) -> Unit){
-        backendService.deleteClub(request).enqueue(object : Callback<ResDeleteClub> {
+        retrofitClient.deleteClub(request).enqueue(object : Callback<ResDeleteClub> {
             override fun onResponse(call: Call<ResDeleteClub>, response: Response<ResDeleteClub>) {
                 if (response.isSuccessful) {
                     val out: ResDeleteClub? = response.body()
@@ -263,7 +203,7 @@ class BackendRepository {
         })
     }
     suspend fun getUserProfileWithUsername(request: InGetUserProfileWithUsername,  onSuccess: (ResGetUserProfileWithUsername?) -> Unit, onError: (String) -> Unit){
-        backendService.getUserProfileWithUsername(request).enqueue(object : Callback<ResGetUserProfileWithUsername> {
+        retrofitClient.getUserProfileWithUsername(request).enqueue(object : Callback<ResGetUserProfileWithUsername> {
             override fun onResponse(call: Call<ResGetUserProfileWithUsername>, response: Response<ResGetUserProfileWithUsername>) {
                 if (response.isSuccessful) {
                     val out: ResGetUserProfileWithUsername? = response.body()
@@ -279,7 +219,7 @@ class BackendRepository {
         })
     }
     suspend fun addFriendConnection(request: InAddFriendConnection,  onSuccess: (ResAddFriendConnection?) -> Unit, onError: (String) -> Unit){
-        backendService.addFriendConnection(request).enqueue(object : Callback<ResAddFriendConnection> {
+        retrofitClient.addFriendConnection(request).enqueue(object : Callback<ResAddFriendConnection> {
             override fun onResponse(call: Call<ResAddFriendConnection>, response: Response<ResAddFriendConnection>) {
                 if (response.isSuccessful) {
                     val out: ResAddFriendConnection? = response.body()
@@ -295,7 +235,7 @@ class BackendRepository {
         })
     }
     suspend fun getFriendConnections(request: InGetFriendConnections,  onSuccess: (ResGetFriendConnections?) -> Unit, onError: (String) -> Unit){
-        backendService.getFriendConnections(request).enqueue(object : Callback<ResGetFriendConnections> {
+        retrofitClient.getFriendConnections(request).enqueue(object : Callback<ResGetFriendConnections> {
             override fun onResponse(call: Call<ResGetFriendConnections>, response: Response<ResGetFriendConnections>) {
                 if (response.isSuccessful) {
                     val out: ResGetFriendConnections? = response.body()
@@ -311,7 +251,7 @@ class BackendRepository {
         })
     }
     suspend fun removeFriend(request: InRemoveFriend,  onSuccess: (ResRemoveFriend?) -> Unit, onError: (String) -> Unit){
-        backendService.removeFriend(request).enqueue(object : Callback<ResRemoveFriend> {
+        retrofitClient.removeFriend(request).enqueue(object : Callback<ResRemoveFriend> {
             override fun onResponse(call: Call<ResRemoveFriend>, response: Response<ResRemoveFriend>) {
                 if (response.isSuccessful) {
                     val out: ResRemoveFriend? = response.body()
@@ -327,7 +267,7 @@ class BackendRepository {
         })
     }
     suspend fun joinClub(request: InJoinClub,  onSuccess: (ResJoinClub?) -> Unit, onError: (String) -> Unit){
-        backendService.joinClub(request).enqueue(object : Callback<ResJoinClub> {
+        retrofitClient.joinClub(request).enqueue(object : Callback<ResJoinClub> {
             override fun onResponse(call: Call<ResJoinClub>, response: Response<ResJoinClub>) {
                 if (response.isSuccessful) {
                     val out: ResJoinClub? = response.body()
@@ -343,7 +283,7 @@ class BackendRepository {
         })
     }
     suspend fun unjoinClub(request: InUnjoinClub,  onSuccess: (ResUnjoinClub?) -> Unit, onError: (String) -> Unit){
-        backendService.unjoinClub(request).enqueue(object : Callback<ResUnjoinClub> {
+        retrofitClient.unjoinClub(request).enqueue(object : Callback<ResUnjoinClub> {
             override fun onResponse(call: Call<ResUnjoinClub>, response: Response<ResUnjoinClub>) {
                 if (response.isSuccessful) {
                     val out: ResUnjoinClub? = response.body()
@@ -359,7 +299,7 @@ class BackendRepository {
         })
     }
     suspend fun getClubMembers(request: InGetClubMembers,  onSuccess: (ResGetClubMembers?) -> Unit, onError: (String) -> Unit){
-        backendService.getClubMembers(request).enqueue(object : Callback<ResGetClubMembers> {
+        retrofitClient.getClubMembers(request).enqueue(object : Callback<ResGetClubMembers> {
             override fun onResponse(call: Call<ResGetClubMembers>, response: Response<ResGetClubMembers>) {
                 if (response.isSuccessful) {
                     val out: ResGetClubMembers? = response.body()
@@ -375,7 +315,7 @@ class BackendRepository {
         })
     }
     suspend fun getUserClubs(request: InGetUserClubs,  onSuccess: (ResGetUserClubs?) -> Unit, onError: (String) -> Unit){
-        backendService.getUserClubs(request).enqueue(object : Callback<ResGetUserClubs> {
+        retrofitClient.getUserClubs(request).enqueue(object : Callback<ResGetUserClubs> {
             override fun onResponse(call: Call<ResGetUserClubs>, response: Response<ResGetUserClubs>) {
                 if (response.isSuccessful) {
                     val out: ResGetUserClubs? = response.body()
@@ -391,7 +331,7 @@ class BackendRepository {
         })
     }
     suspend fun regexSearchClubs(request: InRegexSearchClubs,  onSuccess: (ResRegexSearchClubs?) -> Unit, onError: (String) -> Unit){
-        backendService.regexSearchClubs(request).enqueue(object : Callback<ResRegexSearchClubs> {
+        retrofitClient.regexSearchClubs(request).enqueue(object : Callback<ResRegexSearchClubs> {
             override fun onResponse(call: Call<ResRegexSearchClubs>, response: Response<ResRegexSearchClubs>) {
                 if (response.isSuccessful) {
                     val out: ResRegexSearchClubs? = response.body()
@@ -407,7 +347,7 @@ class BackendRepository {
         })
     }
     suspend fun regexSearchUsers(request: InRegexSearchUsers,  onSuccess: (ResRegexSearchUsers?) -> Unit, onError: (String) -> Unit){
-        backendService.regexSearchUsers(request).enqueue(object : Callback<ResRegexSearchUsers> {
+        retrofitClient.regexSearchUsers(request).enqueue(object : Callback<ResRegexSearchUsers> {
             override fun onResponse(call: Call<ResRegexSearchUsers>, response: Response<ResRegexSearchUsers>) {
                 if (response.isSuccessful) {
                     val out: ResRegexSearchUsers? = response.body()
@@ -418,6 +358,54 @@ class BackendRepository {
             }
 
             override fun onFailure(call: Call<ResRegexSearchUsers>, t: Throwable) {
+                onError(t.message ?: "Network request failed")
+            }
+        })
+    }
+    suspend fun getRandomClubs(request: InGetRandomClubs,  onSuccess: (ResGetRandomClubs?) -> Unit, onError: (String) -> Unit){
+        retrofitClient.getRandomClubs(request).enqueue(object : Callback<ResGetRandomClubs> {
+            override fun onResponse(call: Call<ResGetRandomClubs>, response: Response<ResGetRandomClubs>) {
+                if (response.isSuccessful) {
+                    val out: ResGetRandomClubs? = response.body()
+                    onSuccess(out)
+                } else {
+                    onError(response.errorBody()?.string() ?: "Unknown error occurred")
+                }
+            }
+
+            override fun onFailure(call: Call<ResGetRandomClubs>, t: Throwable) {
+                onError(t.message ?: "Network request failed")
+            }
+        })
+    }
+    suspend fun checkToken(request: InCheckToken,  onSuccess: (ResCheckToken?) -> Unit, onError: (String) -> Unit){
+        retrofitClient.checkToken(request).enqueue(object : Callback<ResCheckToken> {
+            override fun onResponse(call: Call<ResCheckToken>, response: Response<ResCheckToken>) {
+                if (response.isSuccessful) {
+                    val out: ResCheckToken? = response.body()
+                    onSuccess(out)
+                } else {
+                    onError(response.errorBody()?.string() ?: "Unknown error occurred")
+                }
+            }
+
+            override fun onFailure(call: Call<ResCheckToken>, t: Throwable) {
+                onError(t.message ?: "Network request failed")
+            }
+        })
+    }
+    suspend fun unsafeAddFile(request: InUnsafeAddFile,  onSuccess: (ResUnsafeAddFile?) -> Unit, onError: (String) -> Unit){
+        retrofitClient.unsafeAddFile(request).enqueue(object : Callback<ResUnsafeAddFile> {
+            override fun onResponse(call: Call<ResUnsafeAddFile>, response: Response<ResUnsafeAddFile>) {
+                if (response.isSuccessful) {
+                    val out: ResUnsafeAddFile? = response.body()
+                    onSuccess(out)
+                } else {
+                    onError(response.errorBody()?.string() ?: "Unknown error occurred")
+                }
+            }
+
+            override fun onFailure(call: Call<ResUnsafeAddFile>, t: Throwable) {
                 onError(t.message ?: "Network request failed")
             }
         })
